instructpix2pix_backend-main/requirements.txt
python-telegram-bot
diffusers
transformers
accelerate
pillow
requests
torch
fastapi
pydantic
timeout-decorator